backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
pydantic[email-validator]==2.7.4
qrcode[pil]==7.4.2
Pillow>=10.4.0
python-dotenv==1.0.1
slowapi==0.1.9
aiosqlite==0.20.0
httpx==0.27.0
greenlet>=3.0.0
redis>=5.0.0
python-telegram-bot>=21.0
